requirements.txt
streamlit==1.38.0
groq
requests==2.32.3
newspaper3k==0.2.8
python-dotenv==1.0.1
httpx<0.28